requirements.txt
anthropic>=0.40.0
openai>=1.54.0
python-dotenv>=1.0.0
pandas>=2.0.0
pdfplumber>=0.10.0
streamlit>=1.30.0
openpyxl>=3.1.0
python-docx>=1.1.0
xlrd>=2.0.1
requests>=2.31.0
